refactor(packages/viewer): modularize components
