+++ packages/viewer/src/components/default-spinner.tsx
@@ -1,0 +1,14 @@
+import { MutatingDots } from 'react-loader-spinner';
+
+const DefaultSpinner = () => {
+  return (
+    <MutatingDots
+      height="100px"
+      width="100px"
+      color="#fff"
+      secondaryColor="#fff"
+    />
+  );
+};
+
+export default DefaultSpinner;
