+++ packages/viewer/src/components/index.ts
@@ -1,0 +1,2 @@
+export { default as DefaultSpinner } from './default-spinner';
+export { default as SpinnerWrapper } from './spinner-wrapper';
