+++ packages/viewer/src/components/scene/index.ts
@@ -1,0 +1,4 @@
+export { default as SceneModel } from './scene-model';
+export { default as SceneCamera, type CameraProps } from './scene-camera';
+export { default as SceneControls, type ControlsProps } from './scene-controls';
+export { default as SceneGrid, type GridProps } from './scene-grid';
