--- packages/viewer/src/lib/index.ts
@@ -1,1 +1,0 @@
-export { default as VectrealViewer } from './vectreal-viewer';
